fixed favorites workflow, made item-details sharable screen, updated profile screen with more actions buttons
