+++ app/(tabs)/menu/_layout.tsx
@@ -15,18 +15,6 @@ export default function MenuLayout() {
           headerShown: false 
         }} 
       />
-      <Stack.Screen 
-        name="item-details" 
-        options={{
-          headerTitle: 'Item Details',
-          // headerBackTitle: ' ',
-          // headerRight: () => (
-          //   <Ionicons name="heart-outline" size={24} color="#000" style={{ marginRight: 16 }} />
-          // ),
-          headerShown: false,
-        }}
-      />
-      {/* <Stack.Screen name="place-order" options={{ title: 'Place Order' }} /> */}
     </Stack>
   );
 }
