+++ app/item-details/[itemId].tsx
@@ -3,10 +3,19 @@ import { View, Text, Image, StyleSheet, TouchableOpacity } from 'react-native';
 import { useLocalSearchParams, useNavigation, usePathname, useRouter } from 'expo-router';
 import { Button } from 'react-native-paper';
 import { Ionicons } from '@expo/vector-icons';
-import { mockMenuItems } from '../../../mockData';
-import { theme } from '../../../styles/theme';
+import { mockMenuItems } from '../../mockData';
+import { theme } from '../../styles/theme';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import useCustomHeader from '../../../hooks/useCustomHeader';
+import useCustomHeader from '../../hooks/useCustomHeader';
+
+type CartItem = {
+  _id: string;
+  name: string;
+  image: string;
+  description: string;
+  price: number;
+  quantity: number;
+};
 
 export default function ItemDetailsScreen() {
   const pathname = usePathname();
@@ -17,8 +26,8 @@ export default function ItemDetailsScreen() {
   const item = mockMenuItems.find(i => i._id === itemId);
   const [quantity, setQuantity] = useState(1);
   const [isFavorite, setIsFavorite] = useState(false);
-  
-  useCustomHeader({title: "Item Details", showBackButton: false, onBackPress: null, showCartButton: true, onCartPress: ()=>router.push('/cart'), customHeaderOptions: {}});
+
+  useCustomHeader({title: "Item Details", showBackButton: true, onBackPress: null, showCartButton: true, onCartPress: ()=>router.push('/cart'), customHeaderOptions:{}});
 
   useEffect(() => {
     const checkFavoriteStatus = async () => {
@@ -49,14 +58,27 @@ export default function ItemDetailsScreen() {
     setIsFavorite(!isFavorite);
   };
 
-  const addToOrder = () => {
-    // Here you would typically update your order state or send to an API
-    alert(`Added ${quantity} ${item.name}(s) to your order!`);
+  const addToCart = async () => {
+    try {
+      const cartData = await AsyncStorage.getItem('cart');
+      let cart = cartData ? JSON.parse(cartData) : [];
+      const existingItem = cart.find((cartItem: CartItem) => cartItem._id === item._id);
+      
+      if (existingItem) {
+        existingItem.quantity += quantity;
+      } else {
+        cart.push({ ...item, quantity });
+      }
+      
+      await AsyncStorage.setItem('cart', JSON.stringify(cart));
+      alert(`Added ${quantity} ${item.name}(s) to your cart!`);
+      router.back();
+    } catch (error) {
+      console.error('Error adding item to cart:', error);
+      alert('Failed to add item to cart. Please try again.');
+    }
   };
 
-  const placeOrder = () => {
-    router.push({ pathname: '/cart', params: { order: JSON.stringify([{...item, quantity}]) } });
-  };
 
   return (
     <View style={styles.container}>
@@ -84,21 +106,13 @@ export default function ItemDetailsScreen() {
           </TouchableOpacity>
         </View>
         <View style={styles.buttonContainer}>
-          <Button 
+        <Button 
             mode="contained"
-            onPress={addToOrder}
+            onPress={addToCart}
             style={styles.addButton}
             labelStyle={styles.buttonLabel}
           >
-            Add to Order
-          </Button>
-          <Button 
-            mode="contained"
-            onPress={placeOrder}
-            style={styles.orderButton}
-            labelStyle={styles.buttonLabel}
-          >
-            Place Order
+            Add to Cart
           </Button>
         </View>
       </View>
@@ -164,11 +178,6 @@ const styles = StyleSheet.create({
     flex: 1,
     marginRight: theme.spacing.sm,
     backgroundColor: theme.colors.primary,
-  },
-  orderButton: {
-    flex: 1,
-    marginLeft: theme.spacing.sm,
-    backgroundColor: theme.colors.primaryVariant,
   },
   buttonLabel: {
     color: theme.colors.onPrimary,
